requirements.txt
alembic==0.9.6
attrs==17.3.0
certifi==2017.11.5
cffi==1.11.2
chardet==3.0.4
click==6.7
coverage==4.4.2
coveralls==1.2.0
cryptography==1.7.2
docopt==0.6.2
flasgger==0.8.0
Flask==0.12.2
Flask-API==1.0
Flask-Migrate==2.1.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
gunicorn==19.7.1
idna==2.6
itsdangerous==0.24
Jinja2==2.10
jsonschema==2.6.0
Mako==1.0.7
MarkupSafe==1.0
mistune==0.8.3
nose2==0.7.3
pluggy==0.6.0
psycopg2==2.7.3.2
py==1.5.2
pyasn1==0.4.2
pycparser==2.18
PyJWT==1.5.3
pytest==3.3.1
pytest-cov==2.5.1
python-dateutil==2.6.1
python-editor==1.0.3
PyYAML==3.12
requests==2.18.4
six==1.11.0
SQLAlchemy==1.1.15
urllib3==1.22
views==0.3
Werkzeug==0.13